content_builder/pyproject.toml
[project]
name = "content-builder-agent"
version = "0.1.0"
description = "Content Builder microservice agent"
requires-python = ">=3.10,<3.14"
dependencies = [
    "google-adk>=1.16.0,<2.0.0",
    "a2a-sdk~=0.3.9",
    "fastapi~=0.115.8",
    "uvicorn~=0.34.0",
    "google-cloud-logging>=3.12.0,<4.0.0",
    "opentelemetry-exporter-gcp-trace>=1.9.0,<2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
